pyproject.toml
[project]
name = "src"
version = "0.1.0"
authors = [{ name = "Kuriko Iwai", email = "kuriko@versi0n.io" }]
description = "sales prediction"
readme = "README.md"
requires-python = "==3.12.*"
license = {file = "LICENSE"}
keywords = ["machine learning", 'deep learning']
classifiers = ["Programming Language :: Python"]
dependencies = [
    "aws-wsgi>=0.2.7",
    "boto3>=1.39.17",
    "category-encoders>=2.8.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "lightgbm>=4.6.0",
    "numpy>=2.1.3",
    "optuna>=4.4.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "scikit-optimize>=0.10.2",
    "tensorflow>=2.19.0",
    "torch>=2.7.1",
    "waitress>=3.0.2"
]

[project.urls]
Homepage = "https://kuriko.vercel.app"
Repository = "https://github.com/krik8235/ml-sales-prediction"
Issues = "https://github.com/krik8235/ml-sales-prediction/issues"

[project.optional-dependencies]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.uv]
package = true

[tool.uv.workspace]
members = ["kuriko"]

[tool.mypy]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [""]
